requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
email-validator==2.2.0